refactor: Update CircleSquareAnnotationObj to support rotation
Update the CircleSquareAnnotationObj class in circle_square_annotation.ts to support rotation of the square annotation. This change adds a new 'angle' property to the CircleSquareAnnotation interface and modifies the SquareAnnotationObj class to apply the rotation matrix when the angle is provided. This enhancement allows for more flexible positioning and orientation of square annotations in PDF documents.